Group tests and reorganise folders
